型定義まとめる
